Extract pin logic code to new files
